backend/requirement.txt
fastapi
python-binance
uvicorn[standard]
